Moves kubectl config command to staging
