+++ pkg/kubectl/cmd/BUILD
@@ -25,7 +25,6 @@ go_library(
         "//pkg/kubectl/cmd/certificates:go_default_library",
         "//pkg/kubectl/cmd/clusterinfo:go_default_library",
         "//pkg/kubectl/cmd/completion:go_default_library",
-        "//pkg/kubectl/cmd/config:go_default_library",
         "//pkg/kubectl/cmd/convert:go_default_library",
         "//pkg/kubectl/cmd/cp:go_default_library",
         "//pkg/kubectl/cmd/create:go_default_library",
@@ -62,6 +61,7 @@ go_library(
         "//staging/src/k8s.io/cli-runtime/pkg/genericclioptions:go_default_library",
         "//staging/src/k8s.io/client-go/tools/clientcmd:go_default_library",
         "//staging/src/k8s.io/component-base/cli/flag:go_default_library",
+        "//staging/src/k8s.io/kubectl/pkg/cmd/config:go_default_library",
         "//vendor/github.com/spf13/cobra:go_default_library",
         "//vendor/github.com/spf13/pflag:go_default_library",
     ],
