+++ staging/src/k8s.io/kubectl/pkg/cmd/config/delete_cluster.go
