+++ staging/src/k8s.io/kubectl/pkg/cmd/config/delete_cluster_test.go
