+++ staging/src/k8s.io/kubectl/pkg/cmd/config/delete_context.go
