+++ staging/src/k8s.io/kubectl/pkg/cmd/config/view.go
